+++ src/files/ImportSGL.h
@@ -10,5 +10,6 @@
 namespace ImportSGL {
 
 std::vector<Mesh*> importSGL(const std::vector<uint8_t>& buffer, Scene& scene, AngleRenderer& renderer, SculptManager* sculpt = nullptr);
+std::vector<uint8_t> extractThumbnail(const std::vector<uint8_t>& buffer);
 
 } // namespace ImportSGL